anomaly_detection/requirements.txt
kafka-python==2.2.15
pyspark==4.0.0
